+++ FluidSim/Object.cpp
@@ -3,16 +3,16 @@
 #include "Camera.h"
 #include <glm/gtc/matrix_transform.hpp>
 
-const glm::vec3 Object::m_xAxis = glm::vec3(1, 0, 0);
-const glm::vec3 Object::m_yAxis = glm::vec3(0, 1, 0);
-const glm::vec3 Object::m_zAxis = glm::vec3(0, 0, 1);
-const glm::vec3 Object::m_ForwardVector = -m_zAxis;
-const glm::vec3 Object::m_LeftVector = -m_xAxis;
-const glm::vec3 Object::m_BackwardVector = -m_ForwardVector;
-const glm::vec3 Object::m_RightVector = -m_LeftVector;
+const glm::dvec3 Object::m_xAxis = glm::dvec3(1, 0, 0);
+const glm::dvec3 Object::m_yAxis = glm::dvec3(0, 1, 0);
+const glm::dvec3 Object::m_zAxis = glm::dvec3(0, 0, 1);
+const glm::dvec3 Object::m_ForwardVector = -m_zAxis;
+const glm::dvec3 Object::m_LeftVector = -m_xAxis;
+const glm::dvec3 Object::m_BackwardVector = -m_ForwardVector;
+const glm::dvec3 Object::m_RightVector = -m_LeftVector;
 
-glm::vec3 Camera::m_LocalLookDirection = Object::m_ForwardVector;
-glm::vec3 Camera::m_UpVector = Object::m_yAxis;
+glm::dvec3 Camera::m_LocalLookDirection = Object::m_ForwardVector;
+glm::dvec3 Camera::m_UpVector = Object::m_yAxis;
 
 Object::Object(ContainerObject &container) : m_ContainerObjectPtr{&container}
 {
@@ -38,34 +38,34 @@ void Object::deregisterContainerObjectHooks()
 {
 }
 
-void Object::translate(const glm::vec3 & delta)
+void Object::translate(const glm::dvec3 & delta)
 {
-	m_TranslationTransform *= glm::translate(glm::mat4x4(1.0), delta);
+	m_TranslationTransform *= glm::translate(glm::dmat4x4(1.0), delta);
 }
 
-void Object::rotateLocalX(float radians)
+void Object::rotateLocalX(double radians)
 {
 	rotate(radians, m_xAxis);
 }
 
-void Object::rotateLocalY(float radians)
+void Object::rotateLocalY(double radians)
 {
 	rotate(radians, m_yAxis);
 }
 
-void Object::scale(const glm::vec3 & scaleVector)
+void Object::scale(const glm::dvec3 & scaleVector)
 {
 	m_ScaleTransform = glm::scale(m_ScaleTransform, scaleVector);
 }
 
-glm::vec3 Object::position() const
+glm::dvec3 Object::position() const
 {
-	return glm::vec3(m_TranslationTransform[3]);
+	return glm::dvec3(m_TranslationTransform[3]);
 }
 
-void Object::position(const glm::vec3 & position)
+void Object::position(const glm::dvec3 & position)
 {
-	m_TranslationTransform[3] = glm::vec4(position, 1.0);
+	m_TranslationTransform[3] = glm::dvec4(position, 1.0);
 }
 
 void Object::containedIn(ContainerObject & containerObject)
@@ -73,7 +73,7 @@ void Object::containedIn(ContainerObject & containerObject)
 	m_ContainerObjectPtr = &containerObject;
 }
 
-void Object::rotate(float radians, const glm::vec3 & axis)
+void Object::rotate(double radians, const glm::dvec3 & axis)
 {
-	m_RotationTransform *= glm::rotate(glm::mat4x4(1.0), radians, axis);
+	m_RotationTransform *= glm::rotate(glm::dmat4x4(1.0), radians, axis);
 }
