+++ FluidSim/UniformGrid.h
@@ -10,7 +10,7 @@ class UniformGrid : public UniformGridGeometry
 {
 public:
 	UniformGrid(const UniformGrid &that) = delete;
-	UniformGrid(size_t numElements, const glm::vec3 &minCorner, const glm::vec3 &maxCorner)
+	UniformGrid(size_t numElements, const glm::dvec3 &minCorner, const glm::dvec3 &maxCorner)
 		: UniformGridGeometry(numElements, minCorner, maxCorner)
 	{
 		init();
@@ -20,8 +20,8 @@ public:
 	virtual ~UniformGrid() {}
 
 
-	ItemT& atPosition(const glm::vec3 &position) { return privateAtPosition(position); }
-	const ItemT& atPosition(const glm::vec3 &position) const { return privateAtPosition(position); }
+	ItemT& atPosition(const glm::dvec3 &position) { return privateAtPosition(position); }
+	const ItemT& atPosition(const glm::dvec3 &position) const { return privateAtPosition(position); }
 
 	ItemT& atIndices(const glm::uvec3 &indices) { return privateAtIndices(indices); }
 	const ItemT& atIndices(const glm::uvec3 &indices) const { return privateAtIndices(indices); }
@@ -32,14 +32,14 @@ public:
 
 	std::pair<ItemT&, ItemT&> minMaxElement() { return std::minmax_element(m_Points.begin(), m_Points.end()); }
 
-	ItemT interpolate(const glm::vec3 &position) const;
+	ItemT interpolate(const glm::dvec3 &position) const;
 
 private:
 	void init() { m_Points.resize(gridPointCapacity()); }
 
-	ItemT& privateAtPosition(const glm::vec3 &position);
+	ItemT& privateAtPosition(const glm::dvec3 &position);
 	ItemT& privateAtIndices(const glm::uvec3 &indices);
-	glm::uvec3 getResponsiblePointIndices(const glm::vec3 &position) const;
+	glm::uvec3 getResponsiblePointIndices(const glm::dvec3 &position) const;
 	size_t offsetForIndices(const glm::uvec3 &indices) const;
 
 	std::vector<size_t> calculateCellPointsOffsets(size_t startOffset) const;
@@ -50,15 +50,19 @@ private:
 #include "fsmath.h"
 
 template<class ItemT>
-inline ItemT UniformGrid<ItemT>::interpolate(const glm::vec3 & position) const
+inline ItemT UniformGrid<ItemT>::interpolate(const glm::dvec3 & position) const
 {
 	if (fsmath::anyLess(position, minCorner()) || fsmath::anyLess(minCorner() + gridExtent(), position)) {
 		std::runtime_error("UniformGrid::Interpolate: given position is outside of grid");
 	}
 
 	glm::uvec3 responsiblePointIndices = getResponsiblePointIndices(position);
-	glm::vec3 relativePositionNormalized = (position - minCorner()) / cellExtent() - glm::vec3(responsiblePointIndices);
-	glm::vec3 oneMinusRelativePositionNormalized = glm::vec3(1) - relativePositionNormalized;
+	//responsiblePointIndices = glm::clamp(responsiblePointIndices, glm::uvec3(0), pointsAmount()-glm::uvec3(2));
+
+	glm::dvec3 relativePositionNormalized = (position - minCorner()) / cellExtent() - glm::dvec3(responsiblePointIndices);
+	//relativePositionNormalized = glm::clamp(relativePositionNormalized, glm::vec3(0.0), glm::vec3(1.0));
+
+	glm::dvec3 oneMinusRelativePositionNormalized = glm::dvec3(1) - relativePositionNormalized;
 
 	std::vector<size_t> cellPointsOffsets = calculateCellPointsOffsets(offsetForIndices(responsiblePointIndices));
 
@@ -73,7 +77,7 @@ inline ItemT UniformGrid<ItemT>::interpolate(const glm::vec3 & position) const
 }
 
 template<class ItemT>
-inline ItemT& UniformGrid<ItemT>::privateAtPosition(const glm::vec3 &position)
+inline ItemT& UniformGrid<ItemT>::privateAtPosition(const glm::dvec3 &position)
 {
 	glm::uvec3 indices = getResponsiblePointIndices(position);
 	return privateAtIndices(indices);
@@ -86,9 +90,9 @@ inline ItemT& UniformGrid<ItemT>::privateAtIndices(const glm::uvec3 &indices)
 }
 
 template<class ItemT>
-inline glm::uvec3 UniformGrid<ItemT>::getResponsiblePointIndices(const glm::vec3 & position) const
+inline glm::uvec3 UniformGrid<ItemT>::getResponsiblePointIndices(const glm::dvec3 & position) const
 {
-	glm::vec3 relativePosition = position - minCorner();
+	glm::dvec3 relativePosition = position - minCorner();
 	return (relativePosition * cellsPerExtent());
 }
