+++ FluidSim/VortonOctHeapElement.h
@@ -7,7 +7,7 @@
 class VortonOctHeapElement
 {
 public:
-	VortonOctHeapElement(size_t index, const Supervorton &supervorton, VortonOctHeap &owner, const glm::vec3 &extent);
+	VortonOctHeapElement(size_t index, const Supervorton &supervorton, VortonOctHeap &owner, const glm::dvec3 &extent);
 	~VortonOctHeapElement();
 
 	Supervorton& supervorton();
@@ -17,20 +17,20 @@ public:
 
 	bool hasChildren();
 
-	glm::vec3 calculateVelocity(const glm::vec3 &position);
+	glm::dvec3 calculateVelocity(const glm::dvec3 &position);
 
 	std::vector<VortonOctHeapElement*> forwardNeighbors();
 
 private:
-	glm::vec3 calculateVelocityAccurate(const glm::vec3 &position);
-	glm::vec3 calculateVelocityFast(const glm::vec3 &position);
-	glm::vec3 calculateVelocityViaChildren(const glm::vec3 &position);
-	glm::vec3 calculateVelocityViaContainedVortons(const glm::vec3 &position);
+	glm::dvec3 calculateVelocityAccurate(const glm::dvec3 &position);
+	glm::dvec3 calculateVelocityFast(const glm::dvec3 &position);
+	glm::dvec3 calculateVelocityViaChildren(const glm::dvec3 &position);
+	glm::dvec3 calculateVelocityViaContainedVortons(const glm::dvec3 &position);
 
 	VortonOctHeap &m_Owner;
 	size_t m_HeapIndex;
 
-	glm::vec3 m_Extent;
+	glm::dvec3 m_Extent;
 
 	Supervorton m_Supervorton;
 };
